Continued work toward ApplicationLoadBalancer
Removing a target (=== Instance) should (probably) work now, both
individually and as part of the sync cycle. Adding new instances is
going to take a little more work because ApplicationLoadBalancers have a
little bit more infrastructural overhead than Classic load balancers.

This commit also finishes making the change-over from using a string to
represent the load balancer (TrafficSource). There was a datastore that
should needed to move over to the new interface, which meant that there
was a resource that also needed to move over.