Finish first pass at sync
Currently have sync ~ working in the ApplicationLoadBalancer.
At the moment, syncing will
1. Find a target group with the same name as a baragon group, if possible
2. Guarantee that the targets (== instances) that baragon agents think
   that they're on are members of the target group
3. Guarantee that the load balancers that they baragon agents think
   they're talking to can reach all the az's that the baragon agents
   think that they're on
4. De-register targets that don't appear to have a known baragon
   agent associated with them, or which appear to have an uncommunicative
   baragon agent associated with them.

It is limited in that, among other things, it requires that the target
group be named the same as the baragon group. At the moment, the
application load balancer will not support explicit requests to add
or remove targets and requires a certain amount of infrastructure
to already be in place.