Factor out commonality in deregistration checking
I had duplicated some of the functionality in `ElasticLoadBalancer`
w/r/t checking whether an instance / target could be de-registered in
order to keep the scope of the changes that I was making smaller.  This
DRYs it a little bit; makes `ElasticLoadBalancer` take an ID rather than
a model object from the elasticloadbalancingv1, so it can be used from
both the `ApplicationLoadBalancer` and the `ClassicLoadBalancer`.