Remove useless nullity check
This was just a copy-paste relic; when I swapped the order of the try
block and the loop, the error message still checked for the nullity of
the _collection_ and had the _collection_ in the error message. The
nullity check was removed as being unnecessary, and the error message
now refers to the particular group that failed, rather than the
collection of groups.