Debugging issues w/ injecting auth filter
This corrects the injection for the auth filter, which was not
being constructed correctly before, and fixes how it itself gets
injected into Jersey (just using the resource provider, not both
the resource provider and an @Producer annotation). It also makes a
few neatness changes in the pom files.

It still will not work for serving static files.