Implement {register,remove}Instance for ALB
Implement methods to register and remove instances (ie, targets) in the
context of an application load balancer. This works a little bit
differently, semantically, than the equivalent methods for the classic
load balancers. In particular, it treats the `TargetGroup` as the
`TrafficSource`, and registers and deregisters from the `TargetGroup`
rather than from load balancers individually. This means that whereas
before it was possible to register and deregister a single instances
from individual load balancers, now it is only possible to register and
deregister from load balancers on the scope of the target group.

For example, suppose that LB1 and LB2 were talking to Instance.  It
would have been possible to remove Instance from LB2 and leave it
talking to LB1. With ApplicationLoadBalancers, LB1 and LB2 talk to a
target group, so they can't be individually talked to at the instance
level.

It winds up working out the same way, if you consider
ClassicLoadBalancers and TargetGroups to be the equivalent
TrafficSources, rather than ClassicLoadBalancers and
ApplicationLoadBalancers. You just have the target belong to multiple
TargetGroups.