Rely on container URIInfo rather than context
Was injecting the HttpServletContext and relying on that to pull out the
query parameters. This was creating a scoping competition between Guice
and Jersey to inject certain fields, which was a mess. This no longer
injects the request context and relies on a slightly more-raw way of
pulling out query parameters.

I had originally moved away from this because it was barfing over
certain encodings, eg a '%' in the URL that didn't correspond to an
escape character. This was compounded by some earlier issues. It
shouldn't come up now in a serious way because those issues are
resolved, but it will likely still barf in some circumstances.