Upgrade docker-compose file version
Upgrade the docker-compose file version from 1 to 2.1. This isn't the
most recent available file version (which is 3), but the most recent
version is only available on Docker 1.10. The reason for this is that
docker-compose file format version 1 is scheduled for eventual
deprecation (https://docs.docker.com/compose/compose-file/#/version-1),
so it seemed to make sense to try an upgrade. They should be
functionally identical.