Catch failure to find target group by name
I had (incorrectly) assumed that if Amazon couldn't find a target group
with a given name, it would return an empty list rather than throwing an
exception. This catches that exception and logs that it couldn't find a
target group with the given name.

The problem was coming from the fact that when an agent registers and
when it de-registers, it broadcasts the *Load Balancer* that it's in.
The ApplicationLoadBalancer couldn't find a target group with the same
name as the load balancer, which was causing problems. This is
effectively a shim fix for it; a more involved fix that modifies what
the agent sends out and how that response gets filtered and used by the
service is coming shortly.