Correct misuse of traffic source name
In the register method, I had (correctly) used the traffic source name
as the name of the *TargetGroup* that the instance was a part of. In the
sync method, however, I incorrectly used it as the name of the *Load
Balancer* that the instance was ultimately going to be talking to. This
changes it so that TrafficSource name has the same interpretation
throughout the program, namely that of the TargetGroup.