Use default config if set
Attaches the merging config provider to the bootstrapper in the
BaragonService in such a way that if the system property
`baragonDefaultConfiguration` is set, then it will use the config file
at that location as the default, overriding those values with any values
that are deployed with the application.

This also includes some changes to how the tests are set up,
specifically by removing Jukito. I wasn't using parameter injection, and
it turned out to be much easier to create the object mapper inline
rather than trying to inject it.