Correctly grab group/rack attributes from model
The way that it was working was that only the name of the upstream being
grabbed; it was not looking up the group/rack attributes in the model at
all when you attempted to remove a single upstream.  This change uses
the name of the upstream to lookup the actual upstream object in the
model, then posts that request to the server.

The values that weren't `upstream` and `requestId` were the defaults;
they weren't really coming from anywhere, they were just being filled
in.