Filter BaragonGroups from sync by TrafficSource
In particular, the ApplicationLoadBalancer should not attempt to
sync BaragonGroups that don't actually have a `TargetGroup`
associated with them. This just performs a quick check over their
list of `TrafficSource`s to see if any are target groups and, if
so, proceeds with the sync.

This didn't come up in testing because there was only one baragon
group loaded, and by the time we rolled this out the TargetGroup
was already set up. In this case, when there was a larger number
of baragon groups, not all of which had established target groups,
it was attempting to look up the group rather than skipping over
them. Because the Amazon API throws an exception in these cases,
it was propagating issues much further than it needed to.

There was another factor that also came into play here. SLF4j is
overloaded in such a way that when an ERROR happens and an exception
is in the last position, print the error's stack trace rather than
interpolating the error into the accompanying string. This resulted
in a handful of errors with `{}` in the message, which was unhelpful
in debugging the application.