Restore compatibility with former BaragonGroup
Restores compatiblity with the old BaragonGroup serialization format by
changing the name of the added field. That is, earlier changes had
swapped the type of the `sources` field to use `TrafficSource` rather
than strings; this swaps it back and puts `TrafficSource`s on a
`trafficSource` field.

This makes the new version compatible with anything using a prior
version of the source of this project.