AddTrafficSource wired together
The slow part here was that there was an action being called but it was
missing a reducer, so the results of the action weren't reflected in the
state of the application. This bumps that so the `FetchGroup` action
does get mirrored in application state, and everything turns out fine.