Resolve issues made during code review
This principally involved changing where the destructuring was taking
place; in particular, it moved if from the HOC binding into the
beginning of the render method. If also involved adding a loading icon
to handle when the UI had yet to finish the request, and dropping the
global `_` from being included in the vendor bundle.