Attach pagination/search to requests panel
Limit the number of results displayed at a time to 5, and add buttons so
that the user can page through the results. Also add a quick search box
to the table. Neither the paging not the search are nearly as nice as on
the existing system; in particular, the table doesn't have the "Showing
{} from {} (filtered from {})" text on the original, and the search uses
`fuzzy`'s default implementation, which means that the results can seem
a little bit off kilter.