Fix issues that were raised in code review
Involved two major changes: migrating the single-page utils code into
the package utils code, and changing how the request submission was
handled. In particular, w/r/t the latter, it had been the case that
there as an action creator in the `api/services.es6` file that was
responsible for sending remove upstreams requests. That has now been
removed, in favor of delegating to the more general action creator in
`api/requests.es6` that can handle any kind of request, and just
creating the request body at the point where it is triggered.