Rough first pass at ALB information page
At the moment, this is just a list of the target groups and load
balancers. I'm going to have to walk through this to see how much
information I can actually pull out of the endpoints that I have for
this. In any event, at the moment it is just two tables, one for target
groups and one for load balancers. It doesn't have any interactivity or
useful information on it at the moment, but is just a list of some of
the attributes on the data that we are pulling back.