Fill out information TargetGroupDetail page
This just roughly duplicates the information that is available for
classic load balancers, but for the target groups. I'm really going to
need to play with this for a bit to get the presentation right. In
particular, I would like the load-balancers thing to link to the ALB
detail pages, but I have the ARNs, not the load balancer names (which is
what I have been using). This is fine for some things because ARNs are
guaranteed unique and names aren't, but it's less attractive. So I'm
going to have to think about what to do here.