Humanize times on status page
I had left the timestamps in worker lag and ELB worker lag as
milliseconds, rather than providing a useful interpretation of them.
This uses the same format as the exisiting system, where it
appropriately chooses whether to display seconds or milliseconds, and
will not display the time at all if it exceeds a certain threshold.