Reformat ALB, TargetGroups table
Add JSON buttons for both of them. Also change out the information in
the TargetGroups table, removing the ARN and adding the protocol and
port number that the TargetGroup listens on. This also includes a
handful of small style changes like stripping out extra new lines and
having consistent formatting in the router.