Add dialog to edit health check
Add a modal button + dialog to edit a target group's health check.  It
doesn't _do_ anything at the moment except log whatever the input was to
the console because I need to play around with the endpoints on the back
end in order to more accurately map to our use case.