Add link from TargetGroup to associated ALB
Works the same way as links from albs to target groups. In particular,
it just pulls in data and generates a lookup table from ARNs to names,
which is what we're using to identify these things.  It does add the
overhead of an extra AWS call and another HTTP call when loading the
page.