Flatten endpoints
The I had made the endpoints that we were presenting very hierarchical
in order to better represent the organization of Application Load
Balancers internally. For example, a request to delete a rule would have
been made to
`/albs/load-balancers/:load-balancer/listeners/:listener/rules/:rule`.
This does mimic how ALBs are arranged, but neither the `:load-balancer`
nor the `:listener` parameters are actually used internally. While
they're always accumulated on the front-end, they need to be passed
around unnecessarily in order to make the request.  In short, I no
longer really think that the long-form organizational structure is
sufficiently helpful so as to be with keeping.