Use regex to match path parameter
ARNs are for listeners and rules the unique identifiers that I can pass
around in order to make queries and updates to those aspects of the load
balancers. Unfortunately, they can contain unusual characters, eg `/`,
which doesn't play nicely with URLs. My original plan was just to URI
encode those characters.  But it looks like it is decoded by some middle
piece of the pipeline (nginx?) before being passed on, which is causing
Dropwizard to fail to locate the correct resource. This disambiguates
what could possible be the ARN, so Dropwizard should handle it correctly
regardless of how and whether it is URL encoded.