Debug changes w/r/t uri encoding
In particular, I was URI encoding ARNS when they went out (in the wrong
place, as it happens), and then expecting them to be keyed unencoded.
This just takes away all URI encoding, so it shouldn't come up. This
also fixes a variable name that didn't match what I had been expecting
it to be.