Delay refresh of known agents for processing
Slightly janky solution to the problem of correctly refreshing the known
agents list. The UI was attempting to correctly reload the page after
sending the DELETE request, but the GET request was coming back in
between when the DELETE request was accepted and when it was processed.
This artificially creates a slight delay so that the GET request
resolves to the correct data.