Use correct field to list traffic sources
There was a migration on the ELB support branch that meant that a field
in the groups list JSON changed its name. The result was that the groups
list page was using the wrong name for a field and returning all zeros.
The detail page reports the fields value correctly, but the list
doesn't.