Add search bar to groups page
I found the list of all baragon groups on one page to be a little bit
overwhelming. I just cut it down in the same way as for the services
table, where the table now is paginated to 15 items at a time and
includes a search bar to make it a little easier to find the item that
you're looking for.