Show health information for instance on ALBs
Show health details (eg, healthy, unhealthy, initial, ...) for instances
that are attached to an application load balancer. This was relatively
low-hanging fruit because I already had the information available but
was throwing it away on the backend. This includes both the change to
the resource to make this happen and the change to the front-end to
visualize it. Ideally.