Add additional endpoint to get for ELB instances
Adds an additional endpoint to get the health status of instances which
are attached to a classic load balancer. We can probably just use this
request to list the instances in general, rather than trying to merge
the results of the list that we grab from the `/{elbName}` endpoint with
this one.