Neaten up alb detail page
Make the information display on the ALB detail page a little bit more
compact. Removes a lot of the empty space from having a 3 mostly empty
tables dominating the page. The page still doesn't look great, because
I'm working out how to display the listeners, but it looks better than
it did (IMO).