Avoid some calls to AWS API
Previously it was grabbing the `TargetGroup` for each `TrafficSource`
one-at-a-time. This changes that to bundling all of those calls together
by BaragonGroup. This should hopefully remove some unnecessary calls to
AWS. Relevance is that we have increasingly started hitting rate limits
for ApplicationLoadBalancers as we've been rolling out them out to more
and more things.